+++ _24hplusdotnetcore/Services/MC/MCService.cs
@@ -177,7 +177,7 @@ namespace _24hplusdotnetcore.Services.MC
                         dataMC.Request.CompanyTaxNumber = objCustomer.Working.TaxId;
                         dataMC.Request.ShopCode = objCustomer.Loan.SignAddress.Split('-')[0];
                         dataMC.Request.IssuePlace = objCustomer.Loan.SignAddress.Split('-')[1];
-                        dataMC.Request.LoanAmount = Int64.Parse(loanAmount);
+                        dataMC.Request.LoanAmount = Int32.Parse(loanAmount);
                         dataMC.Request.LoanTenor = Int16.Parse(objCustomer.Loan.Term);
                         dataMC.Request.HasInsurance = objCustomer.Loan.BuyInsurance.Equals("true") ? 1 : 0;
